Create  Instances as Return Values
we want to have the method, create an object using the constructor and then return it as the value of the method. Suppose we have a point object and wish to find the midpoint halfway between it and some other target point named OtherPoint which in our case is p2. We would like to write a method, let’s call it halfway, which takes another Point as a parameter and returns the Point that is halfway between the point and the target point it accepts as input.